3/29 user削除機能認証追加

+++ app/controllers/users_controller.rb
@@ -59,9 +59,14 @@ class UsersController < ApplicationController
   end
 
   def destroy
-    user = User.find( params[:id] )
-    user.destroy
-    render json: {message: '削除しました。'}
+    if current_user_check?
+      user = User.find( params[:id] )
+      user.destroy
+      render json: 
+    else
+      render json: {message: '不正な操作です。削除失敗しました。'}
+    end
+   
   end
 
 private
